Create list box with sample content
The rough plan is to have a list of perviously verified files and a big
button in the header bar later.

I am curious how one is meant to keep translation strings up to date.
The "update-po" task seems to be able to add new strings to the po
files.
But when it comes to managing removed/updated strings, it seems to be a
manual process.
Most likely I am misunderstanding something though because it is hard to
imagine, that people keep over 30 translation up to date this way.